requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
opencv-python>=4.5.0
numpy>=1.22.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.5.0
pillow>=8.0.0
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0
roboflow>=0.2.31